requirements.txt
streamlit>=1.41.0
pandas>=2.2.3
requests>=2.32.3
beautifulsoup4>=4.12.3
jinja2>=3.1.4
python-dotenv>=1.0.1
SQLAlchemy>=2.0.36
fastapi>=0.116.0
uvicorn>=0.35.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
